Model is recreated. Logic is done

+++ app/src/main/java/com/example/weather/net/responses/CoordinatesResponse.kt
@@ -1,6 +1,8 @@
 package com.example.weather.net.responses
 
+import java.math.BigDecimal
+
 data class CoordinatesResponse(
-    val lon: Double? = null,
-    val lat: Double? = null
+    val lon: BigDecimal? = null,
+    val lat: BigDecimal? = null
 )
